requirements.txt
gspread
gspread_dataframe
ccxt
pyyaml
pandas
